feat: optimize gelu, randperm, topk ops (#4448)

+++ src/flag_gems/ops/pow.py
@@ -11,7 +11,7 @@ _pow = tl_extra_shim.pow
 @pointwise_dynamic(promotion_methods=[(0, 1, "BOOL_TO_LONG")])
 @triton.jit
 def pow_func(x, exponent):
-    return _pow(x.to(tl.float32), exponent)
+    return _pow(x.to(tl.float32), exponent.to(tl.float32))
 
 
 def pow_tensor_tensor(A, exponent):
@@ -22,7 +22,7 @@ def pow_tensor_tensor(A, exponent):
 @pointwise_dynamic(is_tensor=[True, False], promotion_methods=[(0, 1, "BOOL_TO_LONG")])
 @triton.jit
 def pow_func_tensor_scalar(x, exponent):
-    return _pow(x.to(tl.float32), exponent)
+    return _pow(x.to(tl.float32), exponent.to(tl.float32))
 
 
 def pow_tensor_scalar(A, exponent):
@@ -33,7 +33,7 @@ def pow_tensor_scalar(A, exponent):
 @pointwise_dynamic(is_tensor=[False, True], promotion_methods=[(0, 1, "BOOL_TO_LONG")])
 @triton.jit
 def pow_func_scalar_tensor(x, exponent):
-    return _pow(x.to(tl.float32), exponent)
+    return _pow(x.to(tl.float32), exponent.to(tl.float32))
 
 
 def pow_scalar(A, exponent):
